Miner/requirement.txt
slither-analyzer==0.10.2
tqdm==4.66.1